Moved OrkesMetadataClient
